refactor(identifiers): split government/builder.rs into per-country modules
The primitives/identifiers/government/builder.rs file was 1709 LOC (1533
production + 176 tests) — well over the 800 LOC "split at" threshold and
the worst single-file violation in the audit.

Split into a builder/ directory of 20 per-country submodules. The layout
mirrors the existing Layer 3 wrapper at identifiers/builder/government/
1:1 (per-country files + a single europe.rs bundle + dedicated
test_patterns.rs) so the two layers stay visually parallel.

Pure refactor: GovernmentIdentifierBuilder's public API is unchanged.
Each method body moves verbatim, tests follow their methods into the
matching region file. Largest new file is 187 LOC (india.rs).

This is one slice of the #193 code-health umbrella (file-length
sub-category). The umbrella tracks 95 findings across four categories;
this PR resolves the largest single offender without closing #193.

Refs #193
